Hardening, acceptance suite, test target split
- reset() re-queues stranded records before clearing the cache, so Reset Cache can never discard an unuploaded scan (previously safe only because the startup sync always ran the recovery requeue first)
- 13 new unit tests: queue persistence across relaunch with field fidelity (collector name and collection time survive), offline queue retention, newest-edit-wins dedup (including a user edit replacing a recovery-requeued entry), recovery idempotence after server-date backfill, Reset Cache queue safety, and a 6-case acceptance suite modeled on the confirmed field cases: signed-out retention (43-scan fixture + warning wording), wedged-account retention (19-scan fixture), the half-uploaded exchange (deploy resolves, collect retries, never tag-less), the 95-record recovery sweep with already-landed uploads resolving via skip, the stale-record watermark case, and fleet visibility of recovered records via the daily reconcile
- LocationAppTests.swift (1,814 lines) split into 13 feature-grouped files; shared fixture builders moved to TestFixtures.swift as an XCTestCase extension so call sites are unchanged; project.pbxproj updated accordingly
- Full suite 125/125 green on the iPhone 17 simulator
